More bootstrap ui responsive work.
Chrome on mac and pc, Firefox on man, android browser all work.

ie 9 and 10 don't work. CORS issue?

Chrome on andoid does the query and loads the player but the content
does not play. Need to setup remote debugging.

I don't want to spend time on the ie problem. I had it working once
while debugging but could never get it going again. It should work with
the trusted sites that I setup.